// frontend/src/lib/types.ts
export interface ForecastPoint {
  date: string;
  yhat: number;
  yhat_lower: number;
  yhat_upper: number;
}

export interface TickerForecast {
  ticker: string;
  historical: ForecastPoint[];
  forecast_30d: ForecastPoint;
  forecast_60d: ForecastPoint;
  forecast_90d: ForecastPoint;
  forecast_6m: ForecastPoint;
  forecast_1y: ForecastPoint;
  future_series: ForecastPoint[];
}

export interface ForecastResult {
  portfolio_id: string;
  ticker_forecasts: Record<string, TickerForecast>;
}